# spec/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Business Watchdog core API
  version: 1.0.0
  description: |
    Core contracts for P0 and P1. This document is a specification, not a running API.
    HMAC canonicalization and transport limits are normative in TZ sections 11 and 27.
    UI CRUD, auth, billing, export and credential rotation must be added before their implementation.
    All monetary amounts are decimal strings in minor units. All timestamps are RFC3339 UTC.
servers:
  - url: https://watchdog.example.invalid
...
paths:
...
  /api/v1/tenants/context:
    get:
      tags: [Tenancy]
      operationId: tenantContext
      security:
        - {UserSession: []}
      responses:
        "200":
          description: Resolved tenant context
          content:
            application/json:
              schema:
                type: object
                properties:
                  active_tenant_id: {type: [string, "null"], format: uuid}
                  tenant_context_id: {type: [string, "null"], format: uuid}
                  tenant_context_source: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    IntegrationKeyId:
      type: apiKey
      in: header
      name: X-BW-Key-Id
      description: Key ID plus mandatory raw-body HMAC signature headers; key ID alone never authenticates.
    UserSession:
      type: apiKey
      in: cookie
      name: watchdog_session
      description: Same-origin Sanctum session with CSRF on mutations.
    WorkerToken:
      type: http
      scheme: bearer
      description: Scoped worker credential restricted to private ingress.
...
  responses:
    Problem:
      description: Structured safe problem response; HTTP status determines retry policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        errors: {type: array, items: {type: object}}
        retry_after_seconds: {type: integer, minimum: 0}
...